Split gasman.c into gasman and HPC-GAP parts
